# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/contact-groups/{id}" : {
      "get" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseContactGroupDto"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContactGroupUpdateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseContactGroupDto"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/v1/contact-groups" : {
      "get" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "getByWorkspace",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "isActive",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoContactGroupDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "contact-group-controller" ],
        "operationId" : "create",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ContactGroupCreateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseContactGroupDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/check-policies" : {
      "get" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "getByWorkspace_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "search",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "expectedStatusCode",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "hasLatencyThreshold",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "createdAt"
          }
        }, {
          "name" : "sortDir",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "desc"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePagedResponseDtoCheckPolicyDto"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "check-policy-controller" ],
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CheckPolicyCreateCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCheckPolicyDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/dashboard/stats" : {
      "get" : {
        "tags" : [ "dashboard-controller" ],
        "operationId" : "getWorkspaceStats",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseWorkspaceDashboardStatsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/endpoints/{endpointId}/latency" : {
      "get" : {
        "tags" : [ "dashboard-controller" ],
        "operationId" : "getEndpointLatency",
        "parameters" : [ {
          "name" : "X-Workspace-Id",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "endpointId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListEndpointLatencyDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ContactGroupUpdateCommand" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "emailAddresses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "id", "name" ]
      },
      "ApiResponseContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ContactGroupDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "isActive" : {
            "type" : "boolean"
          },
          "emailAddresses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
      "ApiResponseCheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/CheckPolicyDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "intervalSeconds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "timeoutMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "failureThreshold" : {
            "type" : "integer",
            "format" : "int32"
          },
          "latencyThresholdMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expectedStatusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedResponseBody" : {
            "type" : "string"
          },
          "responseRegex" : {
            "type" : "string"
          }
        }
      },
...
      "ContactGroupCreateCommand" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "emailAddresses" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "name" ]
      },
      "CheckPolicyCreateCommand" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "intervalSeconds" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 5
          },
          "timeoutMillis" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 100
          },
          "retryCount" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          },
          "failureThreshold" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          },
          "latencyThresholdMillis" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedStatusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "expectedResponseBody" : {
            "type" : "string"
          },
          "responseRegex" : {
            "type" : "string"
          }
        },
        "required" : [ "failureThreshold", "intervalSeconds", "name", "retryCount", "timeoutMillis" ]
      },
...
      "ApiResponseWorkspaceDashboardStatsDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/WorkspaceDashboardStatsDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "IncidentSummaryDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endpointId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endpointName" : {
            "type" : "string"
          },
          "startedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "reason" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          }
        }
      },
      "WorkspaceDashboardStatsDto" : {
        "type" : "object",
        "properties" : {
          "workspaceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalEndpoints" : {
            "type" : "integer",
            "format" : "int64"
          },
          "upEndpoints" : {
            "type" : "integer",
            "format" : "int64"
          },
          "downEndpoints" : {
            "type" : "integer",
            "format" : "int64"
          },
          "degradedEndpoints" : {
            "type" : "integer",
            "format" : "int64"
          },
          "openIncidentsCount" : {
            "type" : "integer",
            "format" : "int64"
          },
          "activeIncidents" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IncidentSummaryDto"
            }
          }
        }
      },
      "ApiResponseListEndpointLatencyDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EndpointLatencyDto"
            }
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "EndpointLatencyDto" : {
        "type" : "object",
        "properties" : {
          "checkedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "responseTimeMillis" : {
            "type" : "integer",
            "format" : "int64"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
      "ApiResponsePagedResponseDtoContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoContactGroupDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedResponseDtoContactGroupDto" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ContactGroupDto"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ApiResponsePagedResponseDtoCheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PagedResponseDtoCheckPolicyDto"
          },
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PagedResponseDtoCheckPolicyDto" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CheckPolicyDto"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
    }
  }
}